+++ transformer/transforms/string/email_extract_test.py
@@ -1,9 +1,9 @@
 import unittest
-import emailextract
+import email_extract
 
 class TestStringEmailExtractTransform(unittest.TestCase):
     def test_emailextract(self):
-        transformer = emailextract.StringEmailExtractTransform()
+        transformer = email_extract.StringEmailExtractTransform()
         self.assertEqual(transformer.transform(""), "")
         self.assertEqual(transformer.transform("my email is thomas@hils.us"), "thomas@hils.us")
         self.assertEqual(transformer.transform(None), "")
